refactor(iam): kill constants module
ref: MANAGER-11434
