+++ packages/manager/modules/iam/src/components/actionSelect/ActionTrees.class.js
@@ -44,7 +44,7 @@
  * }} RawAction
  */
 
-import { CUSTOM_RESOURCE_TYPE } from '../../constants';
+import { CUSTOM_RESOURCE_TYPE } from '../../iam.constants';
 
 export default class ActionTrees extends Array {
   /**
