+++ packages/manager/modules/iam/src/components/deleteEntity/deleteEntity.template.html
@@ -31,7 +31,7 @@
                     data-ng-submit="$ctrl.delete()"
                 >
                     <oui-field
-                        data-label="{{:: $ctrl.translate.field | translate: { statement: $root.IAM.DELETE_STATEMENT } }}"
+                        data-label="{{:: $ctrl.translate.field | translate: { statement: $ctrl.DELETE_STATEMENT } }}"
                         data-ng-if="$ctrl.statement"
                     >
                         <input
