+++ packages/manager/modules/iam/src/components/iam/iam.component.js
@@ -1,6 +1,12 @@
-import template from './iam.template.html';
+import { ALERT_ID } from '../../iam.constants';
+
 import './iam.styles.scss';
 
 export default {
-  template,
+  template: `
+    <div class="iam">
+      <div data-ovh-alert="${ALERT_ID}" class="mt-3"></div>
+      <div data-ui-view></div>
+    </div>
+  `,
 };
