--- packages/manager/modules/iam/src/components/iam/iam.template.html
@@ -1,4 +1,0 @@
-<div class="iam">
-    <div data-ovh-alert="{{ $root.IAM.ALERT_ID }}" class="mt-3"></div>
-    <div data-ui-view></div>
-</div>
