+++ packages/manager/modules/iam/src/components/onboarding/onboarding.template.html
@@ -8,7 +8,7 @@
         <a
             class="mt-3 oui-button oui-button_primary oui-button_l text-light"
             data-translate="iam_onboarding_cta_1"
-            ui-sref="{{ :: $root.IAM.ROUTES.CREATE_POLICY }}"
+            ui-sref="iam.createPolicy"
             target="_self"
             rel="noopener"
         >
