+++ packages/manager/modules/iam/src/components/policies/policies.template.html
@@ -4,7 +4,7 @@
         id="{{ $ctrl.datagridId }}"
         data-rows-loader="$ctrl.getItems($config)"
         data-page="{{ $ctrl.cursors.index || 1 }}"
-        data-page-size="{{ $root.IAM.PAGE_SIZE }}"
+        data-page-size="{{ $ctrl.PAGE_SIZE }}"
         data-pagination-mode="arrows"
     >
         <!-- <Columns> -->
