+++ packages/manager/modules/iam/src/components/policy/policy.template.html
@@ -2,7 +2,7 @@
     <oui-header data-heading="{{:: 'iam_policy_header_heading' | translate }}">
         <!-- <Tabs> -->
         <oui-header-tabs>
-            <oui-header-tabs-item data-state="{{ $root.IAM.ROUTES.POLICIES }}"
+            <oui-header-tabs-item data-state="iam.policy.policies"
                 ><span
                     data-translate="iam_policy_header_tabs_item_policies"
                 ></span
@@ -12,8 +12,7 @@
                     data-translate="iam_policy_header_tabs_item_identities"
                 ></span
             ></oui-header-tabs-item>
-            <oui-header-tabs-item
-                data-state="{{ $root.IAM.ROUTES.RESOURCE_GROUPS }}"
+            <oui-header-tabs-item data-state="iam.policy.resourceGroups"
                 ><span
                     data-translate="iam_policy_header_tabs_item_resource_groups"
                 ></span
