+++ packages/manager/modules/iam/src/resolves/guides.resolve.js
@@ -1,4 +1,4 @@
-import { GUIDE } from '../constants';
+import { GUIDE } from '../iam.constants';
 
 // ---------------------------------------------------------------------------------------------------- //
