+++ packages/manager/modules/iam/src/services/apiv2.service.js
@@ -1,4 +1,4 @@
-import { API_ERROR, PAGE_SIZE } from '../constants';
+import { API_ERROR, PAGE_SIZE } from '../iam.constants';
 
 export default class Apiv2Service {
   /* @ngInject */
