+++ packages/manager/modules/iam/src/services/guide.service.js
@@ -1,4 +1,4 @@
-import { GUIDE } from '../constants';
+import { GUIDE } from '../iam.constants';
 
 export default class GuideService {
   /* @ngInject */
